food_server/requirements.txt
flask
flask-cors
recipe-scrapers
ingredient-parser-nlp
inflect
